2_openai/deep_search_mine/requirements.txt
# Core dependencies for OpenAI Agents
openai-agents==0.0.17
openai==1.85.0

# UI framework
gradio==5.33.1

# Environment variables
python-dotenv==1.1.0

# Data validation
pydantic==2.11.5

# HTTP requests
httpx==0.28.1

